+++ packages/common/test/js/bdd/given.js
@@ -1,6 +1,0 @@
-const {Given} = require('@cucumber/cucumber');
-const pageHelper = require('../pageHelper');
-
-Given('page name is {string}', function (name) {
-    pageHelper.pageName(name);
-});